Update on my Update
I totally did update my nav bars this time, I just didn't upload them to the branch. I had uploaded everything to submit, but then got sidetracked and when I came back to submit, I started over because the new pictures was not in the images folder so I erased and started over and forgot to upload the new pages.
Either way, here they are.